Safari-609.1.11.1/LayoutTests/platform/win/fast/events/selectstart-by-drag-expected.txt

This test ensures selectstart is fired exactly once when selecting text by a mouse drag.

Initial state: PASS
Mouse down: FAIL - expected selection to be caret but was range
Moving slightly to the right: FAIL - expected selection to be caret but was range
Moving slightly to the left: FAIL - expected selection to be caret but was range
Moving to the right: PASS
Moving further to the right: PASS
Moving back to the left: FAIL - expected selection to be caret but was range
Moving to the right again: PASS
Mouse down on the right: PASS
Moving to the left: PASS
Done.
